% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_method}
\alias{decompose_method}
\title{The default method for decomposition.}
\usage{
decompose_method()
}
\value{
The default decompose method.
}
\description{
The default method for decomposition.
}
