% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm1.R
\name{lmm1}
\alias{lmm1}
\title{LMM with a single random effect and residual random effect.}
\usage{
lmm1(formula, data, varcov, REML = TRUE, ..., dtol = decompose_tol(),
  dmethod = "evd", verbose = 0)
}
\description{
LMM with a single random effect and residual random effect.
}
