% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wlm.R
\name{wlm}
\alias{wlm}
\title{GLS/WLS using lm.}
\usage{
wlm(formula, data, ..., varcov, tol = decompose_tol())
}
\arguments{
\item{formula}{Formula.}

\item{data}{Data.}

\item{...}{Arguments passed to \code{lm}.}

\item{varcov}{Variance-covariance (relative) matrix of residuals.}

\item{tol}{Tolerance for \code{decompose_varcov_evd}.
The default value is output of \code{decompose_tol} function.}
}
\value{
A modified output of \code{lm}.
}
\description{
GLS/WLS using lm.
}

