% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_tol}
\alias{decompose_tol}
\title{The default tolerance for EVD-based decomposition.}
\usage{
decompose_tol()
}
\value{
The default value.
}
\description{
Tolerance for \code{decompose_varcov_evd} function.
}

