% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_varcov}
\alias{decompose_varcov}
\title{Decomposition of the var-covar matrix.}
\usage{
decompose_varcov(varcov, method = c("evd", "chol_evd", "chol"),
  tol = decompose_tol(), output = c("transform", "all", "evd"))
}
\arguments{
\item{varcov}{Variance-covariance (relative) matrix of residuals.}

\item{method}{One of the following methods for decomposition:
\code{"evd"}, \code{"chol_evd"} or \code{"chol"}.}

\item{tol}{Tolerance for \code{decompose_varcov_evd}.
The default value is output of \code{decompose_evd_tol} function.}

\item{output}{Type of output. The transformtion matrix is returned if \code{"transform"}.
More results of decomposition are returned if \code{"all"}.}
}
\value{
Transformation matrix (in its trasnposed form) needed to pass from WLS to OLS.
}
\description{
Decomposition of the var-covar matrix.
}
