% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlm.R
\name{wlm}
\alias{wlm}
\title{GLS/WLS using lm.}
\usage{
wlm(formula, data, ..., varcov, dtol = decompose_tol(),
  dmethod = decompose_method(), verbose = 0)
}
\arguments{
\item{formula}{Formula.}

\item{data}{Data.}

\item{...}{Arguments passed to \code{lm}.}

\item{varcov}{Variance-covariance (relative) matrix of residuals.}

\item{dtol}{Tolerance for \code{decompose_varcov}.
The default value is output of \code{decompose_tol} function.}

\item{dmethod}{One of the following methods for decomposition:
\code{"evd"}, \code{"chol_evd"} or \code{"chol"}.}
}
\value{
A modified output of \code{lm}.
}
\description{
GLS/WLS using lm.
}
\examples{
# @ http://www.biostat.jhsph.edu/~iruczins/teaching/jf/ch5.pdf

# data
data(longley)
nobs <- nrow(longley)

rho <- 0.64417
V <- diag(nobs)
V <- rho^abs(row(V) - col(V))
  
mod <- wlm(Employed ~ GNP + Population, data = longley, varcov = V)
}

